Split out visible checklist and exclusive checklist option
